libsample/functions.cpp: Fix MSVC compiler warning
shiboken2\tests\libsample\functions.cpp(208): warning C4312: 'reinterpret_cast': conversion from 'int' to 'void *' of greater size

Amends change 6bf1edffe4099456d28b3ceb09c5178b12631556

+++ tests/libsample/functions.cpp
@@ -204,7 +204,7 @@ ClassWithFunctionPointer::ClassWithFunctionPointer()
 
 void ClassWithFunctionPointer::callFunctionPointer(int dummy, void (*fp)(void *))
 {
-    int a = dummy;
+    size_t a = dummy;
     fp(reinterpret_cast<void *>(a));
 }
